requirements.txt
python-dotenv
discord.py
git+https://github.com/Rapptz/asqlite.git
